# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "COLL_APP DOCUMENTATION",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/collections": {
            "get": {
                "tags": [
                    "Collections"
                ],
                "summary": "Gets all collections with filtering possibility",
                "operationId": "collectionsIndex",
                "parameters": [
                    {
                        "name": "isLessThanTargetAmount",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "enum": [
                                "0",
                                "1"
                            ]
                        }
                    },
                    {
                        "name": "remainingAmount[gt]",
                        "in": "query",
                        "description": "Greater then. You need choose only one of remainingAmount filtering parameter!",
                        "required": false,
                        "schema": {
                            "type": "float"
                        }
                    },
                    {
                        "name": "remainingAmount[gte]",
                        "in": "query",
                        "description": "Greater then or equal. You need choose only one of remainingAmount filtering parameter!",
                        "required": false,
                        "schema": {
                            "type": "float"
                        }
                    },
                    {
                        "name": "remainingAmount[lt]",
                        "in": "query",
                        "description": "Less then. You need choose only one of remainingAmount filtering parameter!",
                        "required": false,
                        "schema": {
                            "type": "float"
                        }
                    },
                    {
                        "name": "remainingAmount[lte]",
                        "in": "query",
                        "description": "Less then or equal. You need choose only one of remainingAmount filtering parameter!",
                        "required": false,
                        "schema": {
                            "type": "float"
                        }
                    },
                    {
                        "name": "remainingAmount[eq]",
                        "in": "query",
                        "description": "Equal. You need choose only one of remainingAmount filtering parameter!",
                        "required": false,
                        "schema": {
                            "type": "float"
                        }
                    },
                    {
                        "name": "remainingAmount[ne]",
                        "in": "query",
                        "description": "Not equal. You need choose only one of remainingAmount filtering parameter!",
                        "required": false,
                        "schema": {
                            "type": "float"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": [
                                        {
                                            "id": 1,
                                            "title": "Title",
                                            "description": "Some description.",
                                            "targetAmount": 70000,
                                            "link": "https://www.jakubowski.com"
                                        },
                                        {
                                            "id": 2,
                                            "title": "Title2",
                                            "description": "Some description2.",
                                            "targetAmount": 80000,
                                            "link": "https://www.sdfjakubows.com"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Operation failed."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Collections"
                ],
                "summary": "Creates new collection",
                "operationId": "collectionsStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "NewTitle"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Some new description."
                                    },
                                    "targetAmount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10000
                                    },
                                    "link": {
                                        "type": "string",
                                        "example": "https://www.newcolle.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "collection created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "example": "NewTitle"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Some new description."
                                                },
                                                "targetAmount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 10000
                                                },
                                                "link": {
                                                    "type": "string",
                                                    "example": "https://www.newcolle.com"
                                                },
                                                "contributors": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    },
                                                    "example": []
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Operation failed."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/collections/{id}/contributors": {
            "post": {
                "tags": [
                    "Collections"
                ],
                "summary": "Creates new contributor",
                "operationId": "collectionsContributorStore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "userName": {
                                        "type": "string",
                                        "example": "NewTitle"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 750
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "collection created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "15"
                                                },
                                                "userName": {
                                                    "type": "string",
                                                    "example": "NewTitle"
                                                },
                                                "collectionId": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "amount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 750
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Operation failed."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}